Add support for key rotation / multiple keys
This commit adds very basic support for key rotation. The material
changes are methods to pass in a Collection of keys when validating /
decrypting and some minor refactoring to push encryption logic into the
Key class.

I started to experiment with the idea of a key repository to abstract
away retrieval of the appropriate key depending on whether token
generation or token validation is required. I also experimented with an
interface for key management. In the end, I felt it was more appropriate
for client code to develop their own schemes to do this. However, I
retained those classes in src/test .

In the course of making these changes, I realised I could move some
logic out of static functions. I started to make these changes but
deferred updating the calling code to a future commit.

This commit breaks backward compatibility, so I rev-bumped the minor
revision.

Resolves: #10