Apply PMD rules to all submodules as well as POMs
PMD was misconfigured and was not failing builds on the
fernet-aws-secrets-manager-rotator or fernet-jersey-auth sub modules.
This has been corrected and appropriate improvements have been made
and exclusions have been marked.

It would seem the PMD's Law of Demeter (LoD) rule is over-aggressive.
Although I endorse adherance to the LoD, I found myself having to
suppress many instances of the warning. In the future, I may disable it
in the pmd-ruleset. In addition, the DataFlowAnalysis rule often had to
be ignored. I may also disable this and rely on unit-testing, test
coverage rules, and mutation coverage rules instead.