Clean memory of secret data
This change makes a best effort to clean secret data from memory before
it is made available to another process. There is absolutely no
guarantee that this effort will be successful. In addition, client code
is responsible for securing any secret data it provides to the library
or retrieves from the library.

This change is not a substitute for securing the environment in which
the library is used. Rather it attempts to decrease the surface area for
attacks.

References:
* https://github.com/veorq/cryptocoding#clean-memory-of-secret-data .
* https://www.oracle.com/technetwork/java/seccodeguide-139067.html#2-3