Resolve warnings from newer Mockito and JUnit
Note, there were some issues with using Mockito's
@RunWith(MockitoJUnitRunner.class). It only seemed to manifest when
running the tests in a suite as part of the Maven build. Some errors
only occurred when running either the `install` or `test` goals while
some only occurred when running the `install` goal. This is an example of
one of the build-time errors:

[ERROR] verifyApplyReturnsTokenFromBearerToken(com.macasaet.fernet.jersey.FernetTokenValueParamProviderTest)  Time elapsed: 0 s  <<< ERROR!
org.mockito.exceptions.misusing.RedundantListenerException:

Problems adding Mockito listener.
Listener of type 'MismatchReportingTestListener' has already been added and not removed.
It indicates that previous listener was not removed according to the API.
When you add a listener, don't forget to remove the listener afterwards:
  Mockito.framework().removeListener(myListener);
For more information, see the javadoc for RedundantListenerException class.