Use more realistic examples for Jersey Auth
For simplicity, the previous examples used the Fernet token as,
essentially, a session token without any server-side storage of the
sessions. This is an insecure practice. To make the examples more
realistic and to discourage the use of this library as a replacement for
proper session management, this change does the following:
* The examples of token injection and secret injection are updated to
  persist sessions on the backend.
* A method is provided for revoking sessions prior to the natural
  expiration of the Fernet tokens.